Componentize DemoUserButton and Edit .gitignore
