+++ .gitignore
@@ -14,7 +14,6 @@ node_modules/
 .env.development.local
 .env.test.local
 .env.production.local
-sample.env
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
