+++ client/src/pages/Login/LoginForm/LoginForm.tsx
@@ -5,11 +5,8 @@ import { Formik, FormikHelpers } from 'formik';
 import * as Yup from 'yup';
 import Typography from '@material-ui/core/Typography';
 import { CircularProgress } from '@material-ui/core';
-import login from '../../../helpers/APICalls/login';
-import { useAuth } from '../../../context/useAuthContext';
-import { useSnackBar } from '../../../context/useSnackbarContext';
-import { demoUser } from '../../../mocks/demoUser';
 import useStyles from './useStyles';
+import DemoUserLogin from '../../../components/DemoUserLogin/DemoUserLogin';
 
 interface Props {
   handleSubmit: (
@@ -32,23 +29,6 @@ interface Props {
 
 export default function Login({ handleSubmit }: Props): JSX.Element {
   const classes = useStyles();
-
-  const { updateLoginContext } = useAuth();
-  const { updateSnackBarMessage } = useSnackBar();
-
-  const handleDemoUserLogin = () => {
-    const { email, password } = demoUser;
-    login(email, password).then((data) => {
-      if (data.error) {
-        updateSnackBarMessage(data.error.message);
-      } else if (data.success) {
-        updateLoginContext(data.success);
-      } else {
-        console.error({ data });
-        updateSnackBarMessage('An unexpected error occurred. Please try again');
-      }
-    });
-  };
 
   return (
     <Formik
@@ -110,15 +90,7 @@ export default function Login({ handleSubmit }: Props): JSX.Element {
             placeholder="Enter your password"
           />
           <Box textAlign="center">
-            <Button
-              size="large"
-              className={`${classes.submit} ${classes.demoLoginBtn}`}
-              color="primary"
-              variant="outlined"
-              onClick={handleDemoUserLogin}
-            >
-              Demo User Login
-            </Button>
+            <DemoUserLogin classes={`${classes.submit} ${classes.demoLoginBtn}`} />
             <Button type="submit" size="large" variant="contained" color="primary" className={classes.submit}>
               {isSubmitting ? <CircularProgress style={{ color: 'white' }} /> : 'Login'}
             </Button>
