+++ client/src/pages/SignUp/SignUpForm/SignUpForm.tsx
@@ -5,11 +5,8 @@ import { Formik, FormikHelpers } from 'formik';
 import * as Yup from 'yup';
 import Typography from '@material-ui/core/Typography';
 import { CircularProgress } from '@material-ui/core';
-import login from '../../../helpers/APICalls/login';
-import { useAuth } from '../../../context/useAuthContext';
-import { useSnackBar } from '../../../context/useSnackbarContext';
-import { demoUser } from '../../../mocks/demoUser';
 import useStyles from './useStyles';
+import DemoUserLogin from '../../../components/DemoUserLogin/DemoUserLogin';
 
 interface Props {
   handleSubmit: (
@@ -35,23 +32,6 @@ interface Props {
 
 const SignUpForm = ({ handleSubmit }: Props): JSX.Element => {
   const classes = useStyles();
-
-  const { updateLoginContext } = useAuth();
-  const { updateSnackBarMessage } = useSnackBar();
-
-  const handleDemoUserLogin = () => {
-    const { email, password } = demoUser;
-    login(email, password).then((data) => {
-      if (data.error) {
-        updateSnackBarMessage(data.error.message);
-      } else if (data.success) {
-        updateLoginContext(data.success);
-      } else {
-        console.error({ data });
-        updateSnackBarMessage('An unexpected error occurred. Please try again');
-      }
-    });
-  };
 
   return (
     <Formik
@@ -135,15 +115,7 @@ const SignUpForm = ({ handleSubmit }: Props): JSX.Element => {
           />
 
           <Box textAlign="center">
-            <Button
-              size="large"
-              className={`${classes.submit} ${classes.demoLoginBtn}`}
-              color="primary"
-              variant="outlined"
-              onClick={handleDemoUserLogin}
-            >
-              Demo User Login
-            </Button>
+            <DemoUserLogin classes={`${classes.submit} ${classes.demoLoginBtn}`} />
             <Button type="submit" size="large" variant="contained" color="primary" className={classes.submit}>
               {isSubmitting ? <CircularProgress style={{ color: 'white' }} /> : 'Sign Up'}
             </Button>
